Add support for Announcement Plugin, added action to Cancel a Yes/No Question, bug fixes
Added support for the Announcement Plugin, by creating a variable and speaking its contents. The variable is used in when building an Announcement in the Announcements Plugin when testing.
Fixed a bug when requesting status by properly combining text input fields in the Action Editor screen.
Fixed a bug which caused some speech strings, not to wrap when displaying in the log.