Removed hard coded Voice and Sound files, reworked Logging options, added ability to cancel a yes/no question
Removed hard coded Voice and Sound files, updated the plugin to use the alexa_constants.py file which had to be modify. Removed line feed "\u2028" from all text fields, to allow users to add blank lines for clarity. Changed verbiage to make it more clear what to select if the users does not want an Action Group to execute. Also, made defaults for non-selection of action groups. Added ShowAlerts when validation failed, Fixed a typo in the name of the Action"PlayAudiFileUrl" which prevent validation. Adjust logging so if the 1st line of text doesn't fit, the 2nd line would be used. Adjust the defaults for logging. Added the ability to cancel a yes/no question via scripting.